Implemented LabeledTable into storybook, TabLoadingWrapper load_data doesn't work
