+++ client/src/components/TabLoadingWrapper/TabLoadingWrapper.stories.js
@@ -1,5 +1,7 @@
+import _ from "lodash";
 import React from "react";
 
+import { TabbedContent } from "./TabbedContent";
 import { TabLoadingWrapper } from "./TabLoadingWrapper";
 
 export default {
@@ -8,15 +10,24 @@ export default {
   argTypes: {},
 };
 
-const Template = (props) => {
-  function load_data(data) {
-    return data;
-  }
+const Template = (args) => {
+  // function load_data(data) {
+  //   return data ? data : "Null";
+  // }
 
-  return <TabLoadingWrapper {...props} load_data={load_data} />;
+  const promise = new Promise((resolve, reject) => {
+    const data = "Some data";
+    resolve(data);
+  });
+
+  return (
+    <TabLoadingWrapper
+      {...args}
+      load_data={promise}
+      TabbedContent={<TabbedContent {...args} data={promise} />}
+    />
+  );
 };
 
 export const Basic = Template.bind({});
-Basic.args = {
-  args: {},
-};
+Basic.args = {};
