+++ client/src/components/TabLoadingWrapper/TabbedContent.js
@@ -1,0 +1,14 @@
+import _ from "lodash";
+import React from "react";
+
+const TabbedContent = ({ args, data }) => {
+  // {
+  //   _.map(data, (text, key) => {
+  //     <div key={key}>{text}</div>;
+  //   });
+  // }
+
+  <div>{data}</div>;
+};
+
+export { TabbedContent };
